Slice

+++ src/main/python/model/ffmpeg.py
@@ -72,7 +72,7 @@ UNKNOWN_CHANNEL_LAYOUTS = {
 
 
 def format_biquad(f):
-    return f.format_biquads(False, separator=':')[0]
+    return f.format_biquads(False, separator=':')
 
 
 def get_channel_name(text, channel, channel_count, channel_layout_name='unknown'):
@@ -93,7 +93,7 @@ def get_channel_name(text, channel, channel_count, channel_layout_name='unknown'
         elif channel_layout_name in CHANNEL_LAYOUTS:
             return f"{prefix}{CHANNEL_LAYOUTS[channel_layout_name][channel]}"
         else:
-            return f"{prefix}c{channel+1}"
+            return f"{prefix}c{channel + 1}"
 
 
 def get_next_port():
@@ -127,9 +127,10 @@ class Executor:
         self.__selected_audio_stream_idx = -1
         self.__selected_video_stream_idx = -1
         self.__output_file_name = None
+        self.__output_file_name_overridden = False
         self.__ffmpeg_cmd = None
         self.__ffmpeg_cli = None
-        self.__ffmpeg_filter_complex_script = None
+        self.__filter_complex_script_content = None
         self.__progress_handler = None
         self.__progress_port = get_next_port()
         self.__extractor = None
@@ -139,6 +140,30 @@ class Executor:
         self.__is_remux = signal_model is not None
         self.__signal_model = signal_model
         self.__decimate_fs = int(decimate_fs)
+        self.__start_time_ms = 0
+        self.__end_time_ms = 0
+
+    @property
+    def filter_complex_script_content(self):
+        return self.__filter_complex_script_content
+
+    @property
+    def start_time_ms(self):
+        return self.__start_time_ms
+
+    @start_time_ms.setter
+    def start_time_ms(self, start_time_ms):
+        self.__start_time_ms = start_time_ms
+        self.__calculate_output()
+
+    @property
+    def end_time_ms(self):
+        return self.__end_time_ms
+
+    @end_time_ms.setter
+    def end_time_ms(self, end_time_ms):
+        self.__end_time_ms = end_time_ms
+        self.__calculate_output()
 
     @property
     def channel_to_filter(self):
@@ -240,6 +265,7 @@ class Executor:
     @output_file_name.setter
     def output_file_name(self, output_file_name):
         self.__output_file_name = output_file_name
+        self.__output_file_name_overridden = True
         self.__calculate_ffmpeg_cmd()
 
     @property
@@ -261,7 +287,7 @@ class Executor:
         self.__audio_stream_data = [s for s in self.__probe.get('streams', []) if s['codec_type'] == 'audio']
         self.__video_stream_data = [s for s in self.__probe.get('streams', []) if s['codec_type'] == 'video']
         end = time.time()
-        logger.info(f"Probed {self.file} in {round(end-start, 3)}s")
+        logger.info(f"Probed {self.file} in {round(end - start, 3)}s")
 
     def has_audio(self):
         '''
@@ -420,14 +446,14 @@ class Executor:
     def __calculate_output_file_name(self):
         '''
         Creates a new output file name based on the currently selected stream
-        :return:
         '''
-        stream_idx = str(self.__selected_audio_stream_idx + 1)
-        channel_layout = self.__channel_layout_name
-        output_file_name = f"{Path(self.file).resolve().stem}_s{stream_idx}_{channel_layout}"
-        if self.__mono_mix is True:
-            output_file_name += '_to_mono'
-        self.__output_file_name = output_file_name
+        if self.__output_file_name_overridden is False:
+            stream_idx = str(self.__selected_audio_stream_idx + 1)
+            channel_layout = self.__channel_layout_name
+            output_file_name = f"{Path(self.file).resolve().stem}_s{stream_idx}_{channel_layout}"
+            if self.__mono_mix is True:
+                output_file_name += '_to_mono'
+            self.__output_file_name = output_file_name
 
     def __write_filter_complex(self):
         filts = []
@@ -443,12 +469,13 @@ class Executor:
             filt = f"[a:{self.selected_stream_idx}]pan=1c|c0=c{channel_idx}[{c_name}];"
             filt += f"[{c_name}]aformat=sample_fmts=dbl[{c_name}_{f_idx}];"
             if sig is not None:
-                sig_filter = sig.filter.resample(int(self.__sample_rate))
+                sig_filter = sig.active_filter.resample(int(self.__sample_rate))
                 for f in sig_filter.filters:
-                    filt += f"[{c_name}_{f_idx}]biquad={format_biquad(f)}[{c_name}_{f_idx+1}];"
-                    f_idx += 1
+                    for bq in format_biquad(f):
+                        filt += f"[{c_name}_{f_idx}]biquad={bq}[{c_name}_{f_idx + 1}];"
+                        f_idx += 1
             else:
-                filt += f"[{c_name}_{f_idx}]biquad={format_biquad(Passthrough())}[{c_name}_{f_idx+1}];"
+                filt += f"[{c_name}_{f_idx}]biquad={format_biquad(Passthrough())[0]}[{c_name}_{f_idx + 1}];"
                 f_idx += 1
             filt += f"[{c_name}_{f_idx}]aformat=sample_fmts=s32:sample_rates={self.__sample_rate}"
             filt += f":channel_layouts=mono[{c_name}out]"
@@ -470,13 +497,16 @@ class Executor:
         else:
             merge_filt += '[s0]'
 
-        final_filt = f"{';'.join(filts)};{merge_filt}"
+        new_filt = f"{';'.join(filts)};{merge_filt}"
+        if self.__filter_complex_script_content is None or self.__filter_complex_script_content != new_filt:
+            self.__filter_complex_script_content = new_filt
 
-        if self.__filter_complex_filter is not None:
-            os.remove(self.__filter_complex_filter.name)
-        self.__filter_complex_filter = tempfile.NamedTemporaryFile(delete=False, suffix='.filt', mode='w')
-        print(final_filt, file=self.__filter_complex_filter)
-        self.__filter_complex_filter.close()
+            if self.__filter_complex_filter is not None:
+                os.remove(self.__filter_complex_filter.name)
+            self.__filter_complex_filter = tempfile.NamedTemporaryFile(delete=False, suffix='.filt', mode='w')
+
+            print(self.__filter_complex_script_content, file=self.__filter_complex_filter)
+            self.__filter_complex_filter.close()
         return self.__filter_complex_filter.name.replace('\\', '/')
 
     def __calculate_ffmpeg_cmd(self):
@@ -486,31 +516,22 @@ class Executor:
         else:
             self.__calculate_extract_command(output_file)
 
+    def __calculate_trim_kwargs(self):
+        args = {}
+        if self.start_time_ms > 0:
+            args['ss'] = round(self.start_time_ms / 1000, 3)
+        if self.end_time_ms > 0:
+            args['t'] = round((self.end_time_ms - self.start_time_ms) / 1000, 3)
+        return args
+
     def __calculate_extract_command(self, output_file):
         ''' calculates the command required to extract the audio to the specified output file '''
-        input_stream = ffmpeg.input(self.file)
+        input_stream = ffmpeg.input(self.file, **self.__calculate_trim_kwargs())
         acodec = 'flac' if self.compress_audio else 'pcm_s24le'
         if self.__mono_mix:
-            if self.__selected_video_stream_idx != -1:
-                audio_filter = input_stream['a'].filter('pan', **{'mono|c0': self.__mono_mix_spec})
-                if self.decimate_audio is True:
-                    audio_filter = audio_filter.filter('aresample', str(self.__decimate_fs), resampler='soxr')
-                self.__ffmpeg_cmd = ffmpeg.output(input_stream['v'], audio_filter, output_file,
-                                                  acodec=acodec, vcodec='copy')
-            else:
-                mix = input_stream.filter('pan', **{'mono|c0': self.__mono_mix_spec})
-                if self.decimate_audio is True:
-                    mix = mix.filter('aresample', str(self.__decimate_fs), resampler='soxr')
-                self.__ffmpeg_cmd = mix.output(output_file, acodec=acodec)
+            self.__ffmpeg_cmd = self.__calculate_extract_mono_cmd(acodec, input_stream, output_file)
         else:
-            audio_filter = input_stream[f"a:{self.__selected_audio_stream_idx}"]
-            if self.decimate_audio is True:
-                audio_filter = audio_filter.filter('aresample', str(self.__decimate_fs), resampler='soxr')
-            if self.__selected_video_stream_idx != -1:
-                self.__ffmpeg_cmd = ffmpeg.output(input_stream['v'], audio_filter, output_file, acodec=acodec,
-                                                  vcodec='copy')
-            else:
-                self.__ffmpeg_cmd = audio_filter.output(output_file, acodec=acodec)
+            self.__ffmpeg_cmd = self.__calculate_extract_multi_cmd(acodec, input_stream, output_file)
         if self.progress_handler is not None:
             self.__ffmpeg_cmd = self.__ffmpeg_cmd.global_args('-progress',
                                                               f"udp://127.0.0.1:{self.__progress_port}")
@@ -518,14 +539,41 @@ class Executor:
         self.__ffmpeg_cli = ' '.join(
             [s if s == 'ffmpeg' or s.startswith('-') else f"\"{s}\"" for s in command_args])
 
+    def __calculate_extract_multi_cmd(self, acodec, input_stream, output_file):
+        ''' Calculates an ffmpeg-python cmd for extracting multichannel audio from an input stream. '''
+        audio_filter = input_stream[f"a:{self.__selected_audio_stream_idx}"]
+        if self.decimate_audio is True:
+            audio_filter = audio_filter.filter('aresample', str(self.__decimate_fs), resampler='soxr')
+        if self.__selected_video_stream_idx != -1:
+            return ffmpeg.output(input_stream[f"v:{self.__selected_video_stream_idx}"], audio_filter, output_file,
+                                 acodec=acodec, vcodec='copy')
+        else:
+            return audio_filter.output(output_file, acodec=acodec)
+
+    def __calculate_extract_mono_cmd(self, acodec, input_stream, output_file):
+        ''' Calculates an ffmpeg-python cmd for extracting mono (bass managed) audio from an input stream. '''
+        if self.__selected_video_stream_idx != -1:
+            audio_filter = input_stream[f"a:{self.__selected_audio_stream_idx}"].filter('pan', **{'mono|c0': self.__mono_mix_spec})
+            if self.decimate_audio is True:
+                audio_filter = audio_filter.filter('aresample', str(self.__decimate_fs), resampler='soxr')
+            return ffmpeg.output(input_stream[f"v:{self.__selected_video_stream_idx}"], audio_filter, output_file,
+                                 acodec=acodec, vcodec='copy')
+        else:
+            mix = input_stream[f"a:{self.__selected_audio_stream_idx}"].filter('pan', **{'mono|c0': self.__mono_mix_spec})
+            if self.decimate_audio is True:
+                mix = mix.filter('aresample', str(self.__decimate_fs), resampler='soxr')
+            return mix.output(output_file, acodec=acodec)
+
     def __calculate_remux_cmd(self, output_file):
         ''' calculates the command filter the audio and remux it back into the output file '''
         filename = self.file.replace('\\', '/')
+        trim_args = ' '.join([f"-{key} {value}" for key, value in self.__calculate_trim_kwargs().items()])
+        # TODO only write the filter file when it actually changes
         self.__ffmpeg_cmd = \
-            f"ffmpeg.exe -i \"{filename}\"" \
-            f" -filter_complex_script {self.__write_filter_complex()}"
+            f"ffmpeg.exe {trim_args} -i \"{filename}\"" \
+                f" -filter_complex_script {self.__write_filter_complex()}"
         if self.__selected_video_stream_idx != -1:
-            self.__ffmpeg_cmd += f" -c:v copy -map 0:v:{self.__selected_video_stream_idx}"
+            self.__ffmpeg_cmd += f" -map 0:v:{self.__selected_video_stream_idx} -c:v copy"
         acodec = 'flac' if self.compress_audio else 'pcm_s24le'
         if self.include_original_audio:
             self.__ffmpeg_cmd += f" -map 0:a:{self.__selected_audio_stream_idx}"
